feat: Thumbnail service working, migrate all go projects to docker file, and add a sns fan out system in order to notify services after encode is finished.

+++ metadata/src/main/java/com/shin/metadata/dto/PatchVideoByIdRequest.java
@@ -12,7 +12,7 @@ import java.util.Set;
 public record PatchVideoByIdRequest(
     String title,
     String description,
-    Long duration,
+    Double duration,
     String resolutions,
     String uploadKey,
     String thumbnailUrl,
